requirements.txt
numpy==1.24.3
pandas==2.0.3
Pillow==9.4.0
Pillow==10.2.0
torch==2.1.0
